# pozoristeback/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teatar Maska API",
        "description": "API dokumentacija za seminarski rad",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
...
        "/api/karte": {
            "get": {
                "tags": [
                    "Karte"
                ],
                "summary": "Prikaz svih karata u sistemu",
                "operationId": "ea39ba7b8b3e9c3638b51613a3c7915d",
                "responses": {
                    "200": {
                        "description": "Uspešno učitane karte"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Karte"
                ],
                "summary": "Dodavanje pojedinačne karte",
                "description": "Store metoda - samo za vanredno dodavanje pojedinačne karte.",
                "operationId": "dfcccfd461b3d2439874dff4187cdcdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "izvodjenje_id",
                                    "broj_sedista",
                                    "cena"
                                ],
                                "properties": {
                                    "izvodjenje_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "broj_sedista": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "cena": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 800.5
                                    },
                                    "prodata": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Karta kreirana"
                    },
                    "422": {
                        "description": "Greška u validaciji"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/karte/{id}": {
            "get": {
                "tags": [
                    "Karte"
                ],
                "summary": "Detalji pojedinačne karte",
                "description": "Detalji jedne karte.",
                "operationId": "8a1b1526d092cb06c2564bce0fe14502",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    },
                    "404": {
                        "description": "Karta nije pronađena"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Karte"
                ],
                "summary": "Brisanje karte",
                "description": "Brisanje karte",
                "operationId": "1796d4bf2e97c3eda1c7423dcc9734c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obrisano"
                    },
                    "404": {
                        "description": "Karta nije pronađena"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/karte/izvodjenje/{izvodjenjeId}": {
            "get": {
                "tags": [
                    "Karte"
                ],
                "summary": "Prikaz svih karata za određeno izvođenje",
                "description": "Vraća i slobodne i prodate karte radi prikaza sale u React-u.",
                "operationId": "fb08f9718fb252d49752bd77a027e10c",
                "parameters": [
                    {
                        "name": "izvodjenjeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    },
                    "404": {
                        "description": "Izvođenje nije pronađeno"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Unesite vaš API token koji ste dobili pri prijavi (login) ili registraciji.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}